new auto file to test extended base class
